fix: Make sshfs a core dependency and update docs. (#107)
Using copick with data stored on a remote server with SSH access is a
core part of the workflow now. To avoid confusion when users are
installing for the first time (and potentially forget the extra
dependencies), `sshfs` is moved to the core dependencies. The docs are
updated to reflect this.