fix: Improve warnings when using copick to RELION-friendly DataFrame method (#138)
When provided particles have an identity transformation matrix, the
scipy.spatial.transform.Rotation feature gives:

> UserWarning: Gimbal lock detected. Setting third angle to zero since
it is not possible to uniquely determine all angles.

To prevent user logs from being flooded with this UserWarning when it
really shouldn't be (in the case of an identity transformation matrix),
handle this case separately, but still show warnings for any other
potentially erroneous transformation matrices.

Also improves the wording of the warning of when multiple tomograms of
the same voxel spacing are detected (doesn't include "first", since this
"first" is / can be arbitrary.


I think I messed up the commit history somewhere along the way and so it
appears that there are 12 commits coming in with changes, but there is
actually only one commit with changes that differ from main and the
other commits are from a previous PR (#123) upon which this small PR
builds upon.